Split the oversized arbitrage and bet services (#73)
arbitrage.service.ts (1,096 lines) and bet.service.ts (1,030 lines) each
carried several unrelated responsibilities in one file: detection maths next
to the scan loop next to the analytics queries; bet CRUD next to P&L
aggregation next to pre-placement validation.

Each is now a thin entry point over a directory of single-responsibility
modules. services/arbitrage/ holds types, config (env-derived thresholds and
the per-sport sigma tables), calculator (stake split, stake plan, normal CDF,
middle probability), finder (pure detection over an odds snapshot), alert
(risk scoring and notification), scanner (the scan loop, drift query and
persistence) and query (the reads behind /api/analytics/arbitrage/*).
services/bet/ holds crud, analytics, validation, odds and formatter.

The public API is unchanged. arbitrageService, betService, the ArbitrageService
and BetService classes, and every exported detection helper and type still
resolve from the same two paths, so no route, job, controller or test import
moved. The facades are 123 and 77 lines; the largest module is 429.

Consolidating the reads collapsed the four inline copies of the bet `include`
tree (create, list, get-by-id, post-create refetch) into the two shared
constants the payload types are already derived from.

No behavioural change: getHistory keeps its narrower game selection,
getStatsBySport keeps its unused _where parameter, and the arbitrage
thresholds keep their existing values and env bindings. 774 tests pass, lint
clean, no new type errors.
