+++ dashboard/backend/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Split the two oversized services — arbitrage and bet (issue #73)**: `arbitrage.service.ts` (1,096 lines) and `bet.service.ts` (1,030 lines) each became a thin entry point over a directory of single-responsibility modules. The public API is unchanged: `arbitrageService` / `betService`, the `ArbitrageService` / `BetService` classes, and every exported detection helper and type still resolve from the same two paths, so no route, job, controller or test import moved.
+  - `src/services/arbitrage/` (new): `arbitrage.types.ts` (shared shapes), `arbitrage.config.ts` (env-derived thresholds, per-sport sigma tables), `arbitrage-calculator.service.ts` (stake split, stake plan, normal CDF, middle probability), `arbitrage-finder.service.ts` (pure detection over an odds snapshot, plus a new `detectAll` covering the five detectors the scanner runs), `arbitrage-alert.service.ts` (risk scoring, limit-risk estimate, notification), `arbitrage-scanner.service.ts` (the scan loop, drift query and persistence — the write path), `arbitrage-query.service.ts` (the reads behind `/api/analytics/arbitrage/*` and its MCP mirror)
+  - `src/services/arbitrage.service.ts`: now 123 lines — the module map, the re-exports, and an `ArbitrageService` facade delegating to the scanner and query services. Largest remaining file in the family is 378 lines
+  - `src/services/bet/` (new): `bet-crud.service.ts` (create/read/update/cancel/settle), `bet-analytics.service.ts` (`getStats` and the raw per-sport breakdown), `bet-validation.service.ts` (pre-placement rules, with the game-existence checks lifted into their own `validateGames`), `bet-odds.service.ts` (combined odds and payout, with the parlay and teaser branches split apart), `bet-formatter.ts` (the two Prisma include shapes, their derived payload types, and the response mapping)
+  - `src/services/bet.service.ts`: now 77 lines — a `BetService` facade over the CRUD and analytics modules
+  - The four inline copies of the bet `include` tree (create, list, get-by-id, and the post-create refetch) are now the two shared constants in `bet-formatter.ts`, which are also what the payload types are derived from
+  - No behavioural change: `getHistory` keeps its narrower game selection (no `status`), `getStatsBySport` keeps its unused `_where` parameter, and the arbitrage risk/detection thresholds keep their existing values and env bindings
+
 - **`BaseStatsService` — de-duplicated the seven API-Sports stats services (issue #72)**: The `src/services/api-sports/` family dropped from 2,948 lines to ~2,115 (with the shared code now covered by tests). `syncTeams`, `findTeam`, `getLiveGames`, quota accounting, status mapping, player upsert, and the game/team lookups each have one implementation instead of three to seven.
   - `src/services/api-sports/base-stats.service.ts` (new): abstract `BaseStatsService<TGame>` owning client construction, `getRequestsRemaining`/`hasSufficientQuota`, `getLiveGames`/`getLiveGameIds`, `syncTeams`, `syncTeamStats`, and the shared Prisma helpers (`findTeamByName`, `findTeamByApiId`, `resolveTeam`, `findGameByApiId`, `upsertGameStats`, `upsertPlayer`, `upsertPlayerGameStats`). Each sport supplies only `extractGameId`, `syncGameStats`, and whichever hooks it needs (`liveGameParams`, `teamsParams`, `mapApiTeam`, `filterApiTeams`, `defaultTeamSeason`, `resolveSeasonForDate`)
   - `src/services/api-sports/base-stats.service.ts`: `ScheduledStatsService<TGame>` adds the date-range walk (`syncGamesForDate`) for the three sports with a usable dated `/games` query (NFL, NCAAF, MLB), so the other four don't inherit an API they can't honour
