+++ dashboard/backend/src/services/bet.service.ts
@@ -1,1028 +1,75 @@
-import { prisma } from '../config/database';
-import { logger } from '../config/logger';
-import {
-  CreateBetInput,
-  CreateFutureLegInput,
-  UpdateBetInput,
-  BetFilters,
-  StatsFilters,
-  BetResponse,
-  BetLegResponse,
-  BetLegFutureResponse,
-  PaginatedBets,
-  BetStats,
-  BetStatus
-} from '../types/bet.types';
-import {
-  calculateParlayOdds,
-  calculatePayout,
-  getTeaserOdds,
-  applyTeaserAdjustment,
-  decimalToAmerican
-} from '../utils/odds-calculator';
-import { Decimal } from '@prisma/client/runtime/library';
-import { Prisma } from '@prisma/client';
-import type { Sport } from '../types/betting.types';
-
-// ---------------------------------------------------------------------------
-// Prisma payload types for strongly-typed format helpers
-// ---------------------------------------------------------------------------
-
-const betWithLegsInclude = {
-  legs: {
-    include: {
-      game: {
-        include: {
-          sport: { select: { key: true, name: true } }
-        }
-      }
-    }
-  }
-} as const;
-
-const betWithLegsAndFutureLegsInclude = {
-  legs: {
-    include: {
-      game: {
-        include: {
-          sport: { select: { key: true, name: true } }
-        }
-      }
-    }
-  },
-  futureLegs: {
-    include: {
-      future: {
-        select: {
-          id: true,
-          title: true,
-          season: true,
-          sport: { select: { key: true } }
-        }
-      }
-    }
-  }
-} as const;
-
-type BetWithLegs = Prisma.BetGetPayload<{ include: typeof betWithLegsInclude }>;
-type BetWithLegsAndFutureLegs = Prisma.BetGetPayload<{ include: typeof betWithLegsAndFutureLegsInclude }>;
-type FormattableBet = BetWithLegs | BetWithLegsAndFutureLegs;
-
-type BetLegWithGame = FormattableBet['legs'][number];
-type BetLegFutureWithFuture = BetWithLegsAndFutureLegs['futureLegs'][number];
-
 /**
  * Service for managing bets
- * 
+ *
  * Handles:
  * - Bet creation with validation
  * - Bet updates and cancellation
  * - Bet settlement
  * - Statistics and reporting
+ *
+ * This file is the public entry point. The implementation lives in `./bet/`,
+ * split by responsibility:
+ *
+ *   bet-crud.service.ts        create / read / update / cancel / settle
+ *   bet-analytics.service.ts   P&L, win rate, ROI, per-sport breakdown
+ *   bet-validation.service.ts  pre-placement rules
+ *   bet-odds.service.ts        combined odds and payout for a new bet
+ *   bet-formatter.ts           Prisma include shapes and response mapping
+ */
+
+import {
+  BetFilters,
+  BetResponse,
+  BetStats,
+  CreateBetInput,
+  PaginatedBets,
+  StatsFilters,
+  UpdateBetInput
+} from '../types/bet.types';
+import { betAnalyticsService } from './bet/bet-analytics.service';
+import { betCrudService } from './bet/bet-crud.service';
+
+export { BetAnalyticsService } from './bet/bet-analytics.service';
+export { BetCrudService } from './bet/bet-crud.service';
+export { BetOddsService } from './bet/bet-odds.service';
+export { BetValidationService } from './bet/bet-validation.service';
+
+/**
+ * Thin facade over the bet modules, kept so routes, controllers and jobs
+ * depend on one bet entry point rather than on the internal split.
  */
 export class BetService {
-  /**
-   * Create a new bet
-   */
-  async createBet(data: CreateBetInput, userId?: string): Promise<BetResponse> {
-    logger.info(`Creating ${data.betType} bet: ${data.name}`);
-    
-    // Debug logging for boost
-    if (data.boostedCombinedOdds) {
-      logger.info(`Boost detected in request: ${data.boostedCombinedOdds}`);
-    }
-
-    // Validation
-    await this.validateBetCreation(data);
-
-    // Detect Same Game Parlays (SGP) - multiple legs on same game
-    const gameGroups = new Map<string, typeof data.legs>();
-    data.legs.forEach((leg) => {
-      const gameId = leg.gameId;
-      if (!gameGroups.has(gameId)) {
-        gameGroups.set(gameId, []);
-      }
-      gameGroups.get(gameId)!.push(leg);
-    });
-
-    // Identify which games have SGP (multiple legs)
-    const sgpGames = new Set<string>();
-    gameGroups.forEach((legs, gameId) => {
-      if (legs.length > 1) {
-        sgpGames.add(gameId);
-        logger.info(`SGP detected on game ${gameId}: ${legs.length} legs`);
-      }
-    });
-
-    // Start transaction
-    return await prisma.$transaction(async (tx) => {
-      // Calculate odds and payout
-      const { combinedOdds, potentialPayout } = await this.calculateBetOdds(data);
-
-      // Create bet record
-      const bet = await tx.bet.create({
-        data: {
-          userId: userId || null,
-          name: data.name,
-          betType: data.betType,
-          stake: new Decimal(data.stake),
-          potentialPayout: new Decimal(potentialPayout),
-          status: 'pending',
-          oddsAtPlacement: combinedOdds,
-          teaserPoints: data.teaserPoints ? new Decimal(data.teaserPoints) : null,
-          notes: data.notes || null
-        }
-      });
-
-      // Create bet legs with SGP group IDs
-      for (const legData of data.legs) {
-        // Always use original odds for the odds field, userAdjustedOdds for boosted/adjusted
-        const effectiveLine = legData.userAdjustedLine ?? legData.line;
-
-        let teaserAdjustedLine: Decimal | null = null;
-        if (data.betType === 'teaser' && data.teaserPoints && effectiveLine !== undefined) {
-          const adjusted = applyTeaserAdjustment(
-            effectiveLine,
-            data.teaserPoints,
-            legData.selection
-          );
-          teaserAdjustedLine = new Decimal(adjusted);
-        }
-
-        await tx.betLeg.create({
-          data: {
-            betId: bet.id,
-            gameId: legData.gameId,
-            selectionType: legData.selectionType,
-            selection: legData.selection,
-            teamName: legData.teamName || null,
-            line: effectiveLine !== undefined ? new Decimal(effectiveLine) : null,
-            odds: legData.odds, // Always store original odds here
-            userAdjustedLine: legData.userAdjustedLine ? new Decimal(legData.userAdjustedLine) : null,
-            userAdjustedOdds: legData.userAdjustedOdds || null,
-            teaserAdjustedLine,
-            sgpGroupId: sgpGames.has(legData.gameId) ? legData.gameId : null,
-            bookmaker: (legData as any).bookmaker || null,
-            status: 'pending'
-          }
-        });
-      }
-
-      // Create futures legs if present
-      if (data.futureLegs && data.futureLegs.length > 0) {
-        logger.info(`Creating ${data.futureLegs.length} futures legs for bet ${bet.id}`);
-        for (const futureLeg of data.futureLegs) {
-          const effectiveOdds = futureLeg.userAdjustedOdds || futureLeg.odds;
-          
-          await tx.betLegFuture.create({
-            data: {
-              betId: bet.id,
-              futureId: futureLeg.futureId,
-              outcome: futureLeg.outcome,
-              odds: effectiveOdds,
-              userAdjustedOdds: futureLeg.userAdjustedOdds || null,
-              status: 'pending'
-            }
-          });
-        }
-      }
-
-      const totalLegs = data.legs.length + (data.futureLegs?.length || 0);
-      logger.info(`Created bet ${bet.id} with ${totalLegs} total legs (${data.legs.length} game legs, ${data.futureLegs?.length || 0} futures)${sgpGames.size > 0 ? ` (${sgpGames.size} SGP games)` : ''}`);
-
-      // Fetch complete bet within transaction context
-      const completeBet = await tx.bet.findUnique({
-        where: { id: bet.id },
-        include: {
-          legs: {
-            include: {
-              game: {
-                include: {
-                  sport: {
-                    select: { key: true, name: true }
-                  }
-                }
-              }
-            }
-          },
-          futureLegs: {
-            include: {
-              future: {
-                select: {
-                  id: true,
-                  title: true,
-                  season: true,
-                  sport: {
-                    select: {
-                      key: true
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      });
-
-      if (!completeBet) {
-        throw new Error('Failed to retrieve created bet');
-      }
-
-      return this.formatBet(completeBet);
-    });
+  createBet(data: CreateBetInput, userId?: string): Promise<BetResponse> {
+    return betCrudService.createBet(data, userId);
   }
 
-  /**
-   * Get bets with filters
-   */
-  async getBets(filters: BetFilters = {}): Promise<PaginatedBets> {
-    const limit = filters.limit || 50;
-    const offset = filters.offset || 0;
-
-    // Build where clause
-    const where: Prisma.BetWhereInput = {};
-
-    if (filters.userId) {
-      where.userId = filters.userId;
-    }
-
-    if (filters.status) {
-      where.status = Array.isArray(filters.status)
-        ? { in: filters.status }
-        : filters.status;
-    }
-
-    if (filters.betType) {
-      where.betType = filters.betType;
-    }
-
-    if (filters.startDate || filters.endDate) {
-      where.placedAt = {};
-      if (filters.startDate) {
-        where.placedAt.gte = filters.startDate;
-      }
-      if (filters.endDate) {
-        where.placedAt.lte = filters.endDate;
-      }
-    }
-
-    // Add sport filter through legs
-    if (filters.sportKey) {
-      where.legs = {
-        some: {
-          game: {
-            sport: {
-              key: filters.sportKey
-            }
-          }
-        }
-      };
-    }
-
-    // Get total count
-    const total = await prisma.bet.count({ where });
-
-    // Get bets
-    const bets = await prisma.bet.findMany({
-      where,
-      include: {
-        legs: {
-          include: {
-            game: {
-              include: {
-                sport: {
-                  select: { key: true, name: true }
-                }
-              }
-            }
-          }
-        }
-      },
-      orderBy: { placedAt: 'desc' },
-      take: limit,
-      skip: offset
-    });
-
-    return {
-      bets: bets.map((bet) => this.formatBet(bet)),
-      total,
-      limit,
-      offset
-    };
+  getBets(filters: BetFilters = {}): Promise<PaginatedBets> {
+    return betCrudService.getBets(filters);
   }
 
-  /**
-   * Get bet by ID
-   */
-  async getBetById(id: string, userId?: string): Promise<BetResponse | null> {
-    const bet = await prisma.bet.findFirst({
-      where: {
-        id,
-        ...(userId ? { userId } : {}),
-      },
-      include: {
-        legs: {
-          include: {
-            game: {
-              include: {
-                sport: {
-                  select: { key: true, name: true }
-                }
-              }
-            }
-          }
-        },
-        futureLegs: {
-          include: {
-            future: {
-              select: {
-                id: true,
-                title: true,
-                season: true,
-                sport: {
-                  select: {
-                    key: true
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    });
-
-    return bet ? this.formatBet(bet) : null;
+  getBetById(id: string, userId?: string): Promise<BetResponse | null> {
+    return betCrudService.getBetById(id, userId);
   }
 
-  /**
-   * Update a bet
-   */
-  async updateBet(id: string, data: UpdateBetInput, userId?: string): Promise<BetResponse> {
-    // Get existing bet (scoped by ownership when applicable)
-    const existingBet = await prisma.bet.findFirst({
-      where: {
-        id,
-        ...(userId ? { userId } : {}),
-      },
-      include: {
-        legs: {
-          include: {
-            game: true
-          }
-        }
-      }
-    });
-
-    if (!existingBet) {
-      throw new Error('Bet not found');
-    }
-
-    // Validate can update
-    if (existingBet.status !== 'pending') {
-      throw new Error('Cannot update settled bet');
-    }
-
-    // Check if any games have started
-    const now = new Date();
-    const anyStarted = existingBet.legs.some(
-      leg => leg.game.commenceTime <= now
-    );
-
-    if (anyStarted) {
-      throw new Error('Cannot update bet after games have started');
-    }
-
-    // Prepare update data
-    const updateData: Prisma.BetUpdateManyMutationInput = {};
-
-    if (data.name !== undefined) {
-      updateData.name = data.name;
-    }
-
-    if (data.notes !== undefined) {
-      updateData.notes = data.notes;
-    }
-
-    // Recalculate payout if stake changed
-    if (data.stake !== undefined) {
-      updateData.stake = new Decimal(data.stake);
-
-      if (existingBet.oddsAtPlacement) {
-        const newPayout = calculatePayout(data.stake, existingBet.oddsAtPlacement);
-        updateData.potentialPayout = new Decimal(newPayout);
-      }
-    }
-
-    // SECURITY: Re-assert the ownership filter at update time so the WHERE
-    // clause is atomic with the mutation (closes the TOCTOU between the
-    // findFirst above and the update). updateMany with `count` lets us
-    // detect — and fail loudly on — a concurrent ownership change.
-    const updateResult = await prisma.bet.updateMany({
-      where: {
-        id,
-        ...(userId ? { userId } : {}),
-      },
-      data: updateData,
-    });
-
-    if (updateResult.count === 0) {
-      throw new Error('Bet not found');
-    }
-
-    logger.info(`Updated bet ${id}`);
-
-    const updatedBet = await this.getBetById(id, userId);
-    if (!updatedBet) {
-      throw new Error('Bet not found after update');
-    }
-    return updatedBet;
+  updateBet(id: string, data: UpdateBetInput, userId?: string): Promise<BetResponse> {
+    return betCrudService.updateBet(id, data, userId);
   }
 
-  /**
-   * Cancel a bet
-   */
-  async cancelBet(id: string, force: boolean = false, userId?: string): Promise<void> {
-    // Authorization model:
-    //   - force=false → owner-only: scope by userId so a user can only cancel their own bets.
-    //   - force=true  → admin override (the route gate verifies req.user.isAdmin
-    //                    before calling with force). Admins must be able to delete
-    //                    ANY user's bet, so we drop the userId filter here.
-    const ownershipFilter = !force && userId ? { userId } : {};
-
-    // Get bet with legs
-    const bet = await prisma.bet.findFirst({
-      where: {
-        id,
-        ...ownershipFilter,
-      },
-      include: {
-        legs: {
-          include: {
-            game: true
-          }
-        }
-      }
-    });
-
-    if (!bet) {
-      throw new Error('Bet not found');
-    }
-
-    // If not force delete, validate constraints
-    if (!force) {
-      // Validate can cancel
-      if (bet.status !== 'pending') {
-        throw new Error('Cannot cancel settled bet');
-      }
-
-      // Check if any games have started
-      const now = new Date();
-      const anyStarted = bet.legs.some(
-        leg => leg.game.commenceTime <= now
-      );
-
-      if (anyStarted) {
-        throw new Error('Cannot cancel bet after games have started');
-      }
-    }
-
-    // Delete bet (legs will cascade)
-    await prisma.bet.delete({
-      where: { id: bet.id }
-    });
-
-    logger.info(`${force ? 'Force deleted' : 'Cancelled'} bet ${id}`);
+  cancelBet(id: string, force: boolean = false, userId?: string): Promise<void> {
+    return betCrudService.cancelBet(id, force, userId);
   }
 
-  /**
-   * Manually settle a bet
-   */
-  async settleBet(
+  settleBet(
     id: string,
     status: 'won' | 'lost' | 'push',
     actualPayout?: number,
     userId?: string
   ): Promise<BetResponse> {
-    const bet = await prisma.bet.findFirst({
-      where: {
-        id,
-        ...(userId ? { userId } : {}),
-      },
-      select: { status: true, stake: true, potentialPayout: true }
-    });
-
-    if (!bet) {
-      throw new Error('Bet not found');
-    }
-
-    if (bet.status !== 'pending') {
-      throw new Error('Bet already settled');
-    }
-
-    // Calculate actual payout
-    let payout: Decimal;
-    if (actualPayout !== undefined) {
-      payout = new Decimal(actualPayout);
-    } else if (status === 'won') {
-      payout = bet.potentialPayout || bet.stake;
-    } else if (status === 'push') {
-      payout = bet.stake;
-    } else {
-      payout = new Decimal(0);
-    }
-
-    // Update bet
-    await prisma.bet.update({
-      where: { id },
-      data: {
-        status,
-        actualPayout: payout,
-        settledAt: new Date()
-      }
-    });
-
-    logger.info(`Settled bet ${id}: ${status}`);
-
-    const settledBet = await this.getBetById(id, userId);
-    if (!settledBet) {
-      throw new Error('Bet not found after settlement');
-    }
-    return settledBet;
+    return betCrudService.settleBet(id, status, actualPayout, userId);
   }
 
-  /**
-   * Get betting statistics using database aggregation
-   * Uses groupBy/aggregate instead of loading all rows into memory
-   */
-  async getStats(filters: StatsFilters = {}): Promise<BetStats> {
-    // Build where clause
-    const where: Prisma.BetWhereInput = {};
-
-    if (filters.userId) {
-      where.userId = filters.userId;
-    }
-
-    if (filters.betType) {
-      where.betType = filters.betType;
-    }
-
-    if (filters.startDate || filters.endDate) {
-      where.placedAt = {};
-      if (filters.startDate) {
-        where.placedAt.gte = filters.startDate;
-      }
-      if (filters.endDate) {
-        where.placedAt.lte = filters.endDate;
-      }
-    }
-
-    if (filters.sportKey) {
-      where.legs = {
-        some: {
-          game: {
-            sport: {
-              key: filters.sportKey
-            }
-          }
-        }
-      };
-    }
-
-    // Run all aggregation queries in parallel
-    const [statusGroups, betTypeStatusGroups, sportBreakdown] = await Promise.all([
-      // 1. Group by status: counts + sums
-      prisma.bet.groupBy({
-        by: ['status'],
-        where,
-        _count: true,
-        _sum: {
-          stake: true,
-          actualPayout: true
-        }
-      }),
-
-      // 2. Group by betType + status: counts + sums for byBetType breakdown
-      prisma.bet.groupBy({
-        by: ['betType', 'status'],
-        where,
-        _count: true,
-        _sum: {
-          stake: true,
-          actualPayout: true
-        }
-      }),
-
-      // 3. By sport breakdown using raw SQL (joins through bet_legs → games → sports)
-      this.getStatsBySport(where, filters)
-    ]);
-
-    // Build stats from status groups
-    const stats: BetStats = {
-      totalBets: 0,
-      wonBets: 0,
-      lostBets: 0,
-      pushBets: 0,
-      pendingBets: 0,
-      winRate: 0,
-      totalStaked: 0,
-      totalPayout: 0,
-      netProfit: 0,
-      roi: 0,
-      byBetType: {},
-      bySport: {}
-    };
-
-    for (const group of statusGroups) {
-      const count = group._count;
-      const stake = group._sum.stake?.toNumber() || 0;
-      const payout = group._sum.actualPayout?.toNumber() || 0;
-
-      stats.totalBets += count;
-      stats.totalStaked += stake;
-      stats.totalPayout += payout;
-
-      switch (group.status) {
-        case 'won': stats.wonBets = count; break;
-        case 'lost': stats.lostBets = count; break;
-        case 'push': stats.pushBets = count; break;
-        case 'pending': stats.pendingBets = count; break;
-      }
-    }
-
-    // Build byBetType from betType+status groups
-    for (const group of betTypeStatusGroups) {
-      const betType = group.betType;
-      const stake = group._sum.stake?.toNumber() || 0;
-      const payout = group._sum.actualPayout?.toNumber() || 0;
-
-      if (!stats.byBetType[betType]) {
-        stats.byBetType[betType] = { count: 0, won: 0, netProfit: 0 };
-      }
-      stats.byBetType[betType]!.count += group._count;
-      if (group.status === 'won') stats.byBetType[betType]!.won += group._count;
-      stats.byBetType[betType]!.netProfit += (payout - stake);
-    }
-
-    // Apply sport breakdown
-    stats.bySport = sportBreakdown;
-
-    // Calculate rates
-    stats.netProfit = stats.totalPayout - stats.totalStaked;
-
-    const settledBets = stats.wonBets + stats.lostBets + stats.pushBets;
-    if (settledBets > 0) {
-      stats.winRate = (stats.wonBets / settledBets) * 100;
-    }
-
-    if (stats.totalStaked > 0) {
-      stats.roi = (stats.netProfit / stats.totalStaked) * 100;
-    }
-
-    return stats;
-  }
-
-  /**
-   * Get stats broken down by sport using a raw SQL query.
-   * Joins bet_legs → games → sports to attribute each bet to its first leg's sport.
-   */
-  private async getStatsBySport(
-    _where: Prisma.BetWhereInput,
-    filters: StatsFilters
-  ): Promise<BetStats['bySport']> {
-    // Build parameterized WHERE conditions for the raw query
-    const conditions: string[] = ['1=1'];
-    const params: (string | Date)[] = [];
-    let paramIndex = 1;
-
-    if (filters.userId) {
-      conditions.push(`b.user_id = $${paramIndex++}`);
-      params.push(filters.userId);
-    }
-    if (filters.betType) {
-      conditions.push(`b.bet_type = $${paramIndex++}`);
-      params.push(filters.betType);
-    }
-    if (filters.startDate) {
-      conditions.push(`b.placed_at >= $${paramIndex++}`);
-      params.push(filters.startDate);
-    }
-    if (filters.endDate) {
-      conditions.push(`b.placed_at <= $${paramIndex++}`);
-      params.push(filters.endDate);
-    }
-    if (filters.sportKey) {
-      conditions.push(`s.key = $${paramIndex++}`);
-      params.push(filters.sportKey);
-    }
-
-    const whereClause = conditions.join(' AND ');
-
-    // Use DISTINCT ON to pick first leg per bet, then aggregate by sport
-    const rows = await prisma.$queryRawUnsafe<Array<{
-      sport_key: string;
-      count: bigint;
-      won: bigint;
-      net_profit: number;
-    }>>(
-      `SELECT
-        s.key AS sport_key,
-        COUNT(*)::bigint AS count,
-        COUNT(*) FILTER (WHERE b.status = 'won')::bigint AS won,
-        COALESCE(SUM(COALESCE(b.actual_payout, 0) - b.stake), 0)::float AS net_profit
-      FROM (
-        SELECT DISTINCT ON (bl.bet_id) bl.bet_id, g.sport_id
-        FROM bet_legs bl
-        JOIN games g ON g.id = bl.game_id
-        ORDER BY bl.bet_id, bl.created_at
-      ) first_leg
-      JOIN bets b ON b.id = first_leg.bet_id
-      JOIN sports s ON s.id = first_leg.sport_id
-      WHERE ${whereClause}
-      GROUP BY s.key`,
-      ...params
-    );
-
-    const bySport: BetStats['bySport'] = {};
-    for (const row of rows) {
-      bySport[row.sport_key] = {
-        count: Number(row.count),
-        won: Number(row.won),
-        netProfit: row.net_profit
-      };
-    }
-    return bySport;
-  }
-
-  // ========================================================================
-  // PRIVATE HELPERS
-  // ========================================================================
-
-  /**
-   * Validate bet creation data
-   */
-  private async validateBetCreation(data: CreateBetInput): Promise<void> {
-    // Calculate total legs (game + futures)
-    const totalLegs = data.legs.length + (data.futureLegs?.length || 0);
-
-    // Validate leg count
-    if (totalLegs === 0) {
-      throw new Error('Bet must have at least one leg (game or future)');
-    }
-
-    if (data.betType === 'single' && totalLegs !== 1) {
-      throw new Error('Single bet must have exactly one leg');
-    }
-
-    if (data.betType !== 'single' && totalLegs < 2) {
-      throw new Error('Parlay/teaser must have at least 2 legs');
-    }
-
-    // Validate teaser
-    if (data.betType === 'teaser') {
-      if (!data.teaserPoints) {
-        throw new Error('Teaser must specify points');
-      }
-
-      // Teasers only work with spreads and totals
-      const invalidLegs = data.legs.filter(
-        leg => leg.selectionType === 'moneyline'
-      );
-      if (invalidLegs.length > 0) {
-        throw new Error('Teaser cannot include moneyline selections');
-      }
-    }
-
-    // Validate all games exist and haven't started (only if there are game legs)
-    if (data.legs.length > 0) {
-      const gameIds = data.legs.map(leg => leg.gameId);
-      const uniqueGameIds = [...new Set(gameIds)]; // Deduplicate for validation
-      logger.info(`Validating game IDs: ${JSON.stringify(gameIds)}`);
-      
-      const games = await prisma.game.findMany({
-        where: { id: { in: uniqueGameIds } },
-        select: { 
-          id: true, 
-          commenceTime: true,
-          status: true,
-          homeTeamName: true,
-          awayTeamName: true
-        }
-      });
-
-      logger.info(`Found ${games.length} games out of ${uniqueGameIds.length} unique games requested`);
-      logger.info(`Found game IDs: ${JSON.stringify(games.map(g => g.id))}`);
-      
-      if (games.length !== uniqueGameIds.length) {
-        const foundIds = games.map(g => g.id);
-        const missingIds = uniqueGameIds.filter(id => !foundIds.includes(id));
-        logger.error(`Missing game IDs: ${JSON.stringify(missingIds)}`);
-        throw new Error('One or more games not found');
-      }
-
-      // Block betting on completed or in-progress games, but allow future games
-      const invalidGames = games.filter(game => 
-        game.status === 'final' || game.status === 'in_progress'
-      );
-      if (invalidGames.length > 0) {
-        const statuses = invalidGames.map(g => `${g.homeTeamName} vs ${g.awayTeamName} (${g.status})`).join(', ');
-        throw new Error(`Cannot bet on games that have started or finished: ${statuses}`);
-      }
-    }
-
-    // Validate stake
-    if (data.stake <= 0) {
-      throw new Error('Stake must be positive');
-    }
-  }
-
-  /**
-   * Calculate combined odds and payout
-   */
-  private async calculateBetOdds(data: CreateBetInput): Promise<{
-    combinedOdds: number;
-    potentialPayout: number;
-  }> {
-    const totalLegs = data.legs.length + (data.futureLegs?.length || 0);
-    
-    if (data.betType === 'single') {
-      if (totalLegs !== 1) {
-        throw new Error('Single bet must have exactly one leg');
-      }
-      
-      // Check if it's a game leg or futures leg
-      if (data.legs.length === 1) {
-        const leg = data.legs[0];
-        const odds = leg.odds;
-        const payout = calculatePayout(data.stake, odds);
-        return { combinedOdds: odds, potentialPayout: payout };
-      } else {
-        // Single futures leg
-        const futureLeg = data.futureLegs![0];
-        const odds = futureLeg.odds;
-        const payout = calculatePayout(data.stake, odds);
-        return { combinedOdds: odds, potentialPayout: payout };
-      }
-    }
-
-    if (data.betType === 'parlay') {
-      // Group legs by gameId to detect SGP
-      const gameGroups = new Map<string, typeof data.legs>();
-      data.legs.forEach((leg) => {
-        if (!gameGroups.has(leg.gameId)) {
-          gameGroups.set(leg.gameId, []);
-        }
-        gameGroups.get(leg.gameId)!.push(leg);
-      });
-
-      // Build original odds array for oddsAtPlacement
-      const originalOdds: number[] = [];
-      
-      // Add game leg odds
-      gameGroups.forEach((legs) => {
-        legs.forEach((leg) => {
-          originalOdds.push(leg.odds); // Always original for oddsAtPlacement
-        });
-      });
-      
-      // Add futures leg odds
-      if (data.futureLegs && data.futureLegs.length > 0) {
-        data.futureLegs.forEach((futureLeg) => {
-          originalOdds.push(futureLeg.odds);
-        });
-      }
-
-      // Calculate parlay odds - use ORIGINAL odds for oddsAtPlacement
-      const legsForOriginalCalc = originalOdds.map(odds => ({ odds }));
-      const originalDecimalOdds = calculateParlayOdds(legsForOriginalCalc);
-      const originalAmericanOdds = decimalToAmerican(originalDecimalOdds);
-      
-      // Calculate payout
-      let payout: number;
-      if (data.boostedCombinedOdds) {
-        // Use boosted combined odds for payout calculation
-        payout = calculatePayout(data.stake, data.boostedCombinedOdds);
-        logger.info(`Parlay boost applied: ${originalAmericanOdds} → ${data.boostedCombinedOdds}`);
-      } else {
-        // Use original combined odds for payout
-        payout = calculatePayout(data.stake, originalAmericanOdds);
-      }
-      
-      return { combinedOdds: originalAmericanOdds, potentialPayout: payout };
-    }
-
-    if (data.betType === 'teaser' && data.teaserPoints) {
-      // Get teaser odds from lookup table
-      // Resolve sport from first leg's game record
-      let firstLegSport = 'nfl'; // Default fallback
-      if (data.legs.length > 0) {
-        try {
-          const firstLegGame = await prisma.game.findUnique({
-            where: { id: data.legs[0].gameId },
-            select: { sport: { select: { key: true } } }
-          });
-          if (firstLegGame?.sport?.key) {
-            // Extract sport key (e.g., 'nfl' from 'americanfootball_nfl')
-            const sportKey = firstLegGame.sport.key.split('_').pop() || 'nfl';
-            firstLegSport = sportKey;
-          }
-        } catch (error) {
-          logger.warn('Failed to resolve teaser sport, defaulting to nfl');
-        }
-      }
-      
-      const teaserOdds = getTeaserOdds(
-        data.legs.length,
-        data.teaserPoints,
-        firstLegSport as Sport
-      );
-
-      if (!teaserOdds) {
-        throw new Error('Invalid teaser configuration');
-      }
-
-      const payout = calculatePayout(data.stake, teaserOdds);
-      return { combinedOdds: teaserOdds, potentialPayout: payout };
-    }
-
-    throw new Error('Invalid bet type');
-  }
-
-  /**
-   * Format bet for response.
-   * Prisma returns `string` for all @db.VarChar columns, so narrowing casts to
-   * the union types defined in BetResponse are required here.  The values are
-   * guaranteed valid by input validation at creation time.
-   */
-  private formatBet(bet: FormattableBet): BetResponse {
-    const futureLegs = 'futureLegs' in bet ? bet.futureLegs : undefined;
-    return {
-      id: bet.id,
-      name: bet.name,
-      betType: bet.betType as BetResponse['betType'],
-      stake: bet.stake,
-      potentialPayout: bet.potentialPayout,
-      actualPayout: bet.actualPayout,
-      status: bet.status as BetResponse['status'],
-      oddsAtPlacement: bet.oddsAtPlacement,
-      teaserPoints: bet.teaserPoints,
-      notes: bet.notes,
-      placedAt: bet.placedAt,
-      settledAt: bet.settledAt,
-      legs: bet.legs.map((leg) => this.formatBetLeg(leg)),
-      futureLegs: futureLegs ? futureLegs.map((leg) => this.formatBetLegFuture(leg)) : undefined
-    };
-  }
-
-  /**
-   * Format bet leg for response
-   */
-  private formatBetLeg(leg: BetLegWithGame): BetLegResponse {
-    return {
-      id: leg.id,
-      selectionType: leg.selectionType as BetLegResponse['selectionType'],
-      selection: leg.selection as BetLegResponse['selection'],
-      teamName: leg.teamName,
-      line: leg.line,
-      odds: leg.odds,
-      sgpGroupId: leg.sgpGroupId,
-      userAdjustedLine: leg.userAdjustedLine,
-      userAdjustedOdds: leg.userAdjustedOdds,
-      teaserAdjustedLine: leg.teaserAdjustedLine,
-      status: leg.status as BetLegResponse['status'],
-      game: {
-        id: leg.game.id,
-        externalId: leg.game.externalId,
-        homeTeamName: leg.game.homeTeamName,
-        awayTeamName: leg.game.awayTeamName,
-        commenceTime: leg.game.commenceTime,
-        status: leg.game.status,
-        homeScore: leg.game.homeScore,
-        awayScore: leg.game.awayScore,
-        sport: {
-          key: leg.game.sport.key,
-          name: leg.game.sport.name
-        }
-      }
-    };
-  }
-
-  /**
-   * Format bet leg future for response
-   */
-  private formatBetLegFuture(leg: BetLegFutureWithFuture): BetLegFutureResponse {
-    return {
-      id: leg.id,
-      outcome: leg.outcome,
-      odds: leg.odds,
-      userAdjustedOdds: leg.userAdjustedOdds,
-      status: leg.status as BetLegFutureResponse['status'],
-      future: {
-        id: leg.future.id,
-        title: leg.future.title,
-        sportKey: leg.future.sport.key,
-        season: leg.future.season
-      }
-    };
+  getStats(filters: StatsFilters = {}): Promise<BetStats> {
+    return betAnalyticsService.getStats(filters);
   }
 }
